+++ src/Adjustments/Support/ArrayAdjustmentCollection.php
@@ -88,6 +88,15 @@ class ArrayAdjustmentCollection implements AdjustmentCollectionContract
         }
     }
 
+    public function deleteByType(AdjustmentType $type): void
+    {
+        foreach ($this->items as $key => $item) {
+            if ($type->equals($item->getType())) {
+                unset($this->items[$key]);
+            }
+        }
+    }
+
     public function byType(AdjustmentType $type): AdjustmentCollectionContract
     {
         $result = new self($this->adjustable);